+++ src/features/XIT/BS/BS.vue
@@ -168,7 +168,7 @@ const filteredBases = computed(() => {
       <RadioItem v-model="showInv" horizontal>INV</RadioItem>
       <RadioItem v-model="showWar" horizontal>WAR</RadioItem>
       <div :class="$style.spacer" />
-      <PrunButton primary @click="showBuffer('XIT AGENT')">AGENT</PrunButton>
+      <PrunButton primary @click="showBuffer('XIT AGT')">AGT</PrunButton>
       <PrunButton primary @click="showBuffer('XIT DSP')">DSP</PrunButton>
     </div>
     <table :class="$style.table">